Rework to a pure poll+waitpid loop with timeout support
Also: add a raw transferable instance for `&[u8]`.

Using a signalfd didn't help since it catches SIGCHLDs from the test
runners other threads. To move forward I give up on trying to make
something that plays nice with the test runners custom panic hook: I
suspect now that the reason tests time out when a previous test has
failed is that the forked processes wait for a mutex/lock inherited from
the parent process (and so would never get unlocked).

Another benefit of throwing out signalfd is that its Linux only, so
we're back to a posix implementation.