feat(comms): add firewall, manage pending requests
Add Firewall configuration, and a mpsc channel to explicitly ask for
peer rules and approvals if necessary. Verify for each inbound and
outbound request that they are approved before forwarding them.
Add RequestManager to cache requests that are waiting for approval or a
connection to the remove peer, integrate the PeerConnectionsManager into
the RequestManager. Track all pending events with their current status
in the RequestManager, manage FIFO queue for required Network Actions.