feat(procs): Config of Procedure Input & Output
Enforce explicit configuration of Input- and Output-Data for a
Procedure, instead of always using the preceding procedure's output as
input for the next one.
Change the return value of a Procedure to be a HashMap, that may also
contains output data of intermediate procedure-steps
Allow output to the vault to be only temporary, in which case the record
is revoked after the procedure finished.