feat(proc): Enable remote Execution of Procedures
Implement `Serialize` and `Deserialize` for all types in the procedure
module, to allow execution on a remote peer.
Add wrapper type `PrimitiveProcedure` that wraps all possible primitive
Procedures. `ProcedureStep::then` now collects PrimitiveProcedures in a
Vec, which are sequentially executed when running the Procedure.
Procedure does not need a generic Parameter anymore.