fix(client): remove client mapping
Removed the feature that allows to map the client-path of an inbound
request to a different local client-path.
After rethinking this, I don't see any real use-case for this. Unless
this is really needed it is not worth the extra complexity, especially
as it is very difficult to integrate with the firewall logic.